Don't assume zephyr messages are UTF-8 on the wire
Older clients have a tendency to send latin-1 on the wire, and the
zephyr protocol itself does not require an explicit
encoding. Therefore, make encoding and decoding the responsibility of
the application using PyZephyr.

(PyZephyr will to continue to encode a unicode string to UTF-8 before
sending it if that is handed in)

Reported-by: Arun A. Tharuvai <aatharuv@mit.edu>