Use a shorter default format
This may have practical advantages when the zephyr header is otherwise quite
long. It should have negligible functional impact, since current clients mostly
ignore the field (paying attention to it is a security risk). The particular
default format this uses is a link to a page that will explaining (for any
clients that *do* show it) why they're seeing it and why that's bad, and is the
same format used by barnowl and some other clients.